pyproject.toml
[project]
name = "pishock-cyberpunk-2077"
version = "0.1.0"
description = "Local middleware service to safely map Cyberpunk 2077 events to PiShock actions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "python-pishock>=0.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
middleware-file-ingest = "middleware.file_ingest:main"
middleware-setup-wizard = "middleware.setup_wizard:main"
middleware-demo-event = "middleware.demo_event:main"
middleware-run = "middleware.run:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
testpaths = ["middleware/tests"]
norecursedirs = [".git", ".venv", ".tmp", "__pycache__", "pytest-cache-files-*"]
